requirements/test.txt
# Test dependencies go here.
-r base.txt



coverage==4.2
flake8==3.0.4 # pyup: != 2.6.0
django-test-plus==1.0.15
factory_boy==2.8.1

# pytest
pytest-django==2.9.1
pytest-sugar==0.7.1
